Only export Embed and keep main package at the root of the project
